+++ nana/nana_compiler/src/base/mod.rs
@@ -1,0 +1,45 @@
+#[derive(Clone)]
+pub enum Literal {
+    Int(u64),
+    Float(f64),
+    Str(String),
+    Raw(String),
+}
+
+/// Constructing Ast with From trait
+mod construct {
+    use super::*;
+
+    impl From<u64> for Literal {
+        fn from(i: u64) -> Self {
+            Self::Int (i)
+        }
+    }
+    impl From<f64> for Literal {
+        fn from(f: f64) -> Self {
+            Self::Float (f)
+        }
+    }
+    impl From<String> for Literal {
+        fn from(s: String) -> Self {
+            Self::Str (s)
+        }
+    }
+}
+
+/// Printing Ast.
+mod print {
+    use super::*;
+    use std::fmt;
+
+    impl fmt::Debug for Literal {
+        fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+            match self {
+                Literal::Int(e) => write!(f, "Int({})", e),
+                Literal::Float(e) => write!(f, "Flt({})", e),
+                Literal::Str(e) => write!(f, "Str({})", e),
+                Literal::Raw(e) => write!(f, "Raw({})", e),
+            }
+        }
+    }
+}
