+++ nana/nana_compiler/src/lib.rs
@@ -1,3 +1,4 @@
+mod base;
 mod external;
 mod resolve;
 use external::nana_ast;
